Refactor_2024.12.07.00.28_CheckItemColumnコンポーネントのリファクタリング
